Tap water capacity target (#7)
* refactor(translations): update tap water capacity naming

Both the "current tap water capacity" and the "tap water capacity target" shared the same name.
This commit differentiates them by renaming the target to "Hot water capacity target" in English
and "Varmvattenkapacitet (mål)" in Swedish.

* fix(tap_water_capacity): Update tap water capacity unit to be same as target

target is in "number of people" whereas the api returns "number of half-people"

* Correct naming of metric

---------

Co-authored-by: Marcus Nordström <nordstrom.i.marcus@gmail.com>
Co-authored-by: Per Osbäck <perosb@users.noreply.github.com>